Refactor /data endpoints /data/add_review -> /dorms/review /data/profile    -> /user/profile
