+++ python/ratemydorm/__init__.py
@@ -26,8 +26,6 @@ def create_app(test_config=None):
     from ratemydorm.routes import status
     from ratemydorm.routes import user
     from ratemydorm.routes import searchpage
-    from ratemydorm.routes import data
-    from ratemydorm.routes import test_routes
     from ratemydorm.routes import dorms
 
     # Register routes
@@ -36,8 +34,6 @@ def create_app(test_config=None):
     app.register_blueprint(status.bp)
     app.register_blueprint(user.bp)
     app.register_blueprint(searchpage.bp)
-    app.register_blueprint(data.bp)
-    app.register_blueprint(test_routes.bp)
     app.register_blueprint(dorms.bp)
 
     logging.basicConfig(format='%(asctime)s-%(levelname)s: %(message)s',
